refactor rpc calls 1
